+++ tableglobal.h
@@ -19,7 +19,7 @@ public:
 		STDMETHOD(SaveValue)(BSTR TableName, BSTR ValueName, VARIANT Value);
 		STDMETHOD(StopSound)(BSTR Sound);
 		STDMETHOD(AddObject)(BSTR Name, IDispatch *pdisp);
-		STDMETHOD(get_GetPlayerHWnd)(/*[out, retval]*/ long *pVal);
+		STDMETHOD(get_GetPlayerHWnd)(/*[out, retval]*/ size_t *pVal);
 		STDMETHOD(get_UserDirectory)(/*[out, retval]*/ BSTR *pVal);
 		STDMETHOD(get_StartGameKey)(/*[out, retval]*/ long *pVal);
 		STDMETHOD(EndMusic)();
